requirements.txt
pandas
vaderSentiment
streamlit
matplotlib
river
feedparser
yfinance
plotly
torch
transformers
aiohttp
tensorflow
tf-keras
pytest
python-dotenv